+++ src/app/app.module.ts
@@ -1,15 +1,19 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { MatCardModule } from '@angular/material/card';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { FirstComponent } from './first/first.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
+import { MatCardModule } from '@angular/material/card';
+import { MatButtonModule } from '@angular/material/button';
+
+import { ProductComponent } from './product/product.component';
+import { FirstComponent } from './first/first.component';
+
 @NgModule({
-    declarations: [AppComponent, FirstComponent],
-    imports: [BrowserModule, AppRoutingModule, BrowserAnimationsModule, MatCardModule],
+    declarations: [AppComponent, FirstComponent, ProductComponent],
+    imports: [BrowserModule, AppRoutingModule, BrowserAnimationsModule, MatButtonModule, MatCardModule],
     providers: [],
     bootstrap: [AppComponent],
 })