+++ changelog.md
@@ -3,4 +3,5 @@
 1. Установлен CLI  
 2. Настроены бызовые линтинги/форматеры  
 2. Создан проект и залит на Github.   
-3. Создан First Component 
+3. Создан First Component
+4. Добавлен Product Component